Move OneContactConfigScreen to :core:ui
