+++ app/src/main/kotlin/com/blaxsoftware/directcallwidget/ui/ContactCard.kt
@@ -42,7 +42,7 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.blaxsoftware.directcallwidget.R
 import com.blaxsoftware.directcallwidget.ui.components.Picture
-import com.blaxsoftware.directcallwidget.ui.theme.PortraitCardStyle
+import dev.ferp.dcw.core.ui.theme.PortraitCardStyle
 
 @Composable
 fun ContactCard(
