+++ app/src/main/kotlin/com/blaxsoftware/directcallwidget/ui/components/VerticalPlaceholder.kt
@@ -38,7 +38,7 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.blaxsoftware.directcallwidget.R
-import com.blaxsoftware.directcallwidget.ui.theme.PortraitCardStyle
+import dev.ferp.dcw.core.ui.theme.PortraitCardStyle
 
 @Composable
 fun DcwVerticalPlaceholder(
