+++ core/ui/src/main/res/drawable/baseline_add_photo_alternate_24.xml
