+++ core/ui/src/main/res/values-es/strings.xml
@@ -17,5 +17,6 @@
   -->
 
 <resources>
-    <string name="settings">Settings</string>
+    <string name="pick_image">Elegir imagen</string>
+    <string name="clear_image">Borrar imagen</string>
 </resources>
