+++ core/ui/src/main/res/values/strings.xml
@@ -17,5 +17,6 @@
   -->
 
 <resources>
-    <string name="settings">Settings</string>
+    <string name="pick_image">Pick image</string>
+    <string name="clear_image">Clear image</string>
 </resources>
